MVC extension

+++ src/main/java/com/learning/utils/DictionaryRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 // This will be AUTO IMPLEMENTED by Spring into a Bean called dictionaryRepository
 // CRUD refers Create, Read, Update, Delete
 
-public interface DictionaryRepository extends CrudRepository<Dictionary, Long> {
+public interface DictionaryRepository extends CrudRepository<Dictionary, Integer> {
 
 }
